requirements.txt
pyqt5
pyqtchart
plyer
qtawesome
PyYAML
